The existing EnwikiContentSource and DocMaker classes in lucene benchmark package do not parse the contributor of the article which I do need. Do hack this, I've ripped these two classes from lu cene org.apache.lucene.benchmark.byTask.feeds package into my own package and am using these custom classes directly in my code. This way, I'm able to utilitze the existing code in bechmark pack age, at the same time customize the xml parsing / lucene document creation to match my own needs.
In scope of this commit I create a clone of existing EnwikiContentSource and DocMaker classes and integrate them with my WikiIndexer. In it's current state it's runnable and will create lucene index files in a given output folder. This index does not contain the contributor and the wiki text is indexed as plain text without tokenizing. Will be working on that next.